add comments

+++ python/gigl/distributed/dist_link_prediction_dataset.py
@@ -765,7 +765,15 @@ def _append_non_split_node_ids(
     if test_node_ids.numel():
         node_ids_to_get_max.append(test_node_ids)
     max_node_id = int(max(n.max().item() for n in node_ids_to_get_max)) + 1
-    split_counts = torch.bincount(train_node_ids, minlength=max_node_id).clamp(max=255)
+    # We cast to uint8 here to save some memory.
+    # We also use clamp to avoid overflow to 0,
+    # Without clamp, and if we have 255 nodes in a split,
+    # we will asume we have no nodes in that bucket, which is incorrect.
+    split_counts = (
+        torch.bincount(train_node_ids, minlength=max_node_id)
+        .to(torch.uint8)
+        .clamp(max=255)
+    )
     split_counts.add_(
         torch.bincount(val_node_ids, minlength=max_node_id).clamp(max=255)
     )
